Move base54toUint8Array to the token salting file
