--- utils/encryption/base64ToUint8Array.ts
@@ -1,8 +1,0 @@
-export default function base64ToUint8Array(base64String: string): Uint8Array {
-    const binaryString = atob(base64String);
-    const byteArray = new Uint8Array(binaryString.length);
-    for (let i = 0; i < binaryString.length; i++) {
-        byteArray[i] = binaryString.charCodeAt(i);
-    }
-    return byteArray;
-}
