+++ utils/encryption/tokenSalting.ts
@@ -1,4 +1,11 @@
-import base64ToUint8Array from "./base64ToUint8Array";
+function base64ToUint8Array(base64String: string): Uint8Array {
+    const binaryString = atob(base64String);
+    const byteArray = new Uint8Array(binaryString.length);
+    for (let i = 0; i < binaryString.length; i++) {
+        byteArray[i] = binaryString.charCodeAt(i);
+    }
+    return byteArray;
+}
 
 export function encrypt(plaintext: string): string {
     if(!process.env.NEXT_PUBLIC_ENCRYPTION_KEY) {
